backend/requirements.txt
langchain-groq
langchain-core
streamlit
python-dotenv
fastapi
uvicorn
pandas
python-multipart